WIP - started adding separate class for data handling
