+++ brainles_aurora/inferer/inferer.py
@@ -8,7 +8,6 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Dict, List
 
-import monai
 import nibabel as nib
 import numpy as np
 import torch
@@ -17,24 +16,13 @@ from brainles_aurora.inferer import (
     AuroraInfererConfig,
     BaseConfig,
     DataMode,
-    InferenceMode,
-    Output,
     Device,
+    Output,
 )
+from brainles_aurora.inferer.data import DataHandler
 from brainles_aurora.utils import download_model_weights, remove_path_suffixes
-from monai.data import list_data_collate
 from monai.inferers import SlidingWindowInferer
 from monai.networks.nets import BasicUNet
-from monai.transforms import (
-    Compose,
-    EnsureChannelFirstd,
-    Lambdad,
-    LoadImageD,
-    RandGaussianNoised,
-    ScaleIntensityRangePercentilesd,
-    ToTensord,
-)
-from torch.utils.data import DataLoader
 
 logger = logging.getLogger(__name__)
 
@@ -137,155 +125,8 @@ class AuroraInferer(AbstractInferer):
         logger.info(f"Initialized {self.__class__.__name__} with config: {self.config}")
         self.device = self._configure_device()
 
-        self.validated_images = None
         self.inference_mode = None
         self.input_mode = None
-
-    def _validate_images(
-        self,
-        t1: str | Path | np.ndarray | None = None,
-        t1c: str | Path | np.ndarray | None = None,
-        t2: str | Path | np.ndarray | None = None,
-        fla: str | Path | np.ndarray | None = None,
-    ) -> List[np.ndarray | None] | List[Path | None]:
-        """Validate the input images. \n
-        Verify that the input images exist (for paths) and are all of the same type (NumPy or Nifti).
-        If the input is a numpy array, the input mode is set to DataMode.NUMPY, otherwise to DataMode.NIFTI_FILE.
-
-        Args:
-            t1 (str | Path | np.ndarray | None, optional): T1 modality. Defaults to None.
-            t1c (str | Path | np.ndarray | None, optional): T1C modality. Defaults to None.
-            t2 (str | Path | np.ndarray | None, optional): T2 modality. Defaults to None.
-            fla (str | Path | np.ndarray | None, optional): FLAIR modality. Defaults to None.
-
-        Returns:
-            List[np.ndarray | None] | List[Path | None]: List of validated images.
-        """
-
-        def _validate_image(
-            data: str | Path | np.ndarray | None,
-        ) -> np.ndarray | Path | None:
-            if data is None:
-                return None
-            if isinstance(data, np.ndarray):
-                self.input_mode = DataMode.NUMPY
-                return data.astype(np.float32)
-            if not os.path.exists(data):
-                raise FileNotFoundError(f"File {data} not found")
-            if not (data.endswith(".nii.gz") or data.endswith(".nii")):
-                raise ValueError(
-                    f"File {data} must be a nifti file with extension .nii or .nii.gz"
-                )
-            self.input_mode = DataMode.NIFTI_FILE
-            return Path(data).absolute()
-
-        images = [
-            _validate_image(img)
-            for img in [
-                t1,
-                t1c,
-                t2,
-                fla,
-            ]
-        ]
-
-        not_none_images = [img for img in images if img is not None]
-        assert len(not_none_images) > 0, "No input images provided"
-        # make sure all inputs have the same type
-        unique_types = set(map(type, not_none_images))
-        assert (
-            len(unique_types) == 1
-        ), f"All passed images must be of the same type! Received {unique_types}. Accepted Input types: {list(DataMode)}"
-
-        logger.info(
-            f"Successfully validated input images. Input mode: {self.input_mode}"
-        )
-        return images
-
-    def _determine_inference_mode(
-        self, images: List[np.ndarray | None] | List[Path | None]
-    ) -> InferenceMode:
-        """Determine the inference mode based on the provided images.
-
-        Args:
-            images (List[np.ndarray | None] | List[Path | None]): List of validated images.
-
-        Raises:
-            NotImplementedError: If no model is implemented for the combination of input images.
-        Returns:
-            InferenceMode: Inference mode based on the combination of input images.
-        """
-        _t1, _t1c, _t2, _fla = [img is not None for img in images]
-        logger.info(f"Received files: T1: {_t1}, T1C: {_t1c}, T2: {_t2}, FLAIR: {_fla}")
-
-        # check if files are given in a valid combination that has an existing model implementation
-        mode = IMGS_TO_MODE_DICT.get((_t1, _t1c, _t2, _fla), None)
-
-        if mode is None:
-            raise NotImplementedError(
-                "No model implemented for this combination of images"
-            )
-
-        logger.info(f"Inference mode: {mode}")
-        return mode
-
-    def _get_data_loader(self) -> torch.utils.data.DataLoader:
-        """Get the data loader for inference.
-
-        Returns:
-            torch.utils.data.DataLoader: Data loader for inference.
-        """
-        # init transforms
-        transforms = [
-            (
-                LoadImageD(keys=["images"])
-                if self.input_mode == DataMode.NIFTI_FILE
-                else None
-            ),
-            (
-                EnsureChannelFirstd(keys="images")
-                if (
-                    len(self._get_not_none_files()) == 1
-                    and self.input_mode == DataMode.NIFTI_FILE
-                )
-                else None
-            ),
-            Lambdad(["images"], np.nan_to_num),
-            ScaleIntensityRangePercentilesd(
-                keys="images",
-                lower=0.5,
-                upper=99.5,
-                b_min=0,
-                b_max=1,
-                clip=True,
-                relative=False,
-                channel_wise=True,
-            ),
-            ToTensord(keys=["images"]),
-        ]
-        # Filter None transforms
-        transforms = list(filter(None, transforms))
-        inference_transforms = Compose(transforms)
-
-        # Initialize data dictionary
-        data = {
-            "images": self._get_not_none_files(),
-        }
-
-        # init dataset and dataloader
-        infererence_ds = monai.data.Dataset(
-            data=[data],
-            transform=inference_transforms,
-        )
-
-        data_loader = DataLoader(
-            infererence_ds,
-            batch_size=1,
-            num_workers=self.config.workers,
-            collate_fn=list_data_collate,
-            shuffle=False,
-        )
-        return data_loader
 
     def _get_model(self) -> torch.nn.Module:
         """Get the Aurora model based on the inference mode.
@@ -549,10 +390,15 @@ class AuroraInferer(AbstractInferer):
 
         # check inputs and get mode , if mode == prev mode => run inference, else load new model
         prev_mode = self.inference_mode
-        self.validated_images = self._validate_images(t1=t1, t1c=t1c, t2=t2, fla=fla)
-        self.inference_mode = self._determine_inference_mode(
-            images=self.validated_images
+        data_handler = DataHandler(config=self.config)
+        validated_images = data_handler.validate_images(t1=t1, t1c=t1c, t2=t2, fla=fla)
+        self.inference_mode = data_handler.determine_inference_mode(
+            images=validated_images
         )
+
+        logger.info("Setting up Dataloader")
+        self.data_loader = self._get_data_loader()
+
         if prev_mode != self.inference_mode:
             logger.info("No loaded compatible model found. Loading Model and weights")
             self.model = self._get_model()
@@ -561,8 +407,6 @@ class AuroraInferer(AbstractInferer):
                 f"Same inference mode {self.inference_mode} as previous infer call. Re-using loaded model"
             )
         # self.model.eval()
-        logger.info("Setting up Dataloader")
-        self.data_loader = self._get_data_loader()
 
         # setup output file paths
         self.output_file_mapping = {
@@ -571,7 +415,6 @@ class AuroraInferer(AbstractInferer):
             Output.METASTASIS_NETWORK: metastasis_unbinarized_floats_file,
         }
 
-        ########
         logger.info(f"Running inference on device := {self.device}")
         out = self._sliding_window_inference()
         logger.info(f"Finished inference {os.linesep}")
